+++ lib/dashboard/orders/models/order_status_update_response_model.dart
@@ -267,7 +267,7 @@ class ProductDetail {
   double? productDiscount;
   double? productDiscountedValue;
   int? productQuantity;
-  int? addedCartQuantity;
+  double? addedCartQuantity;
   bool? isReturnable;
   bool? isPerishable;
   int? productHsnCode;
@@ -278,8 +278,8 @@ class ProductDetail {
   int? v;
   int? taxableValue;
   double? productTaxValue;
-  int? productSubTotal;
-  int? productGrandTotal;
+  double? productSubTotal;
+  double? productGrandTotal;
 
   ProductDetail({
     this.productCategory,
@@ -334,7 +334,7 @@ class ProductDetail {
     productDiscount: json["productDiscount"].toDouble(),
     productDiscountedValue: json["productDiscountedValue"].toDouble(),
     productQuantity: json["productQuantity"],
-    addedCartQuantity: json["addedCartQuantity"],
+    addedCartQuantity: json["addedCartQuantity"].toDouble(),
     isReturnable: json["isReturnable"],
     isPerishable: json["isPerishable"],
     productHsnCode: json["productHsnCode"],
@@ -345,8 +345,8 @@ class ProductDetail {
     v: json["__v"],
     taxableValue: json["taxableValue"],
     productTaxValue: json["productTaxValue"]?.toDouble(),
-    productSubTotal: json["productSubTotal"],
-    productGrandTotal: json["productGrandTotal"],
+    productSubTotal: json["productSubTotal"].toDouble(),
+    productGrandTotal: json["productGrandTotal"].toDouble(),
   );
 
   Map<String, dynamic> toJson() => {
